src/ui.c: fix sprintf() parameter call
    ui.c:812:37: warning: format '%lx' expects argument of type 'long unsigned int', but argument 3 has type 'unsigned int' [-Wformat=]
      812 |         (void)sprintf(buf, "cr2 %08lx cr3 %08llx [%05x|%s|%s]",
          |                                 ~~~~^
          |                                     |
          |                                     long unsigned int
          |                                 %08x
      813 |                       cr3cr2.pfla,
          |                       ~~~~~~~~~~~
          |                             |
          |                             unsigned int

+++ src/ui.c
@@ -810,7 +810,7 @@ static char *iarLine(unsigned i)
 	tmpreg = arGet(viewPid, CR3_CR2_ID);
 	cr3cr2 = *(IAcr3cr2Ptr)&tmpreg;
 	(void)sprintf(buf, "cr2 %08llx cr3 %08llx [%05x|%s|%s]",
-		      cr3cr2.pfla,
+		      (unsigned long long)cr3cr2.pfla,
 		      (unsigned long long)BitfX(*(DWORD*)&cr3cr2, 32, 32),
 		      (unsigned int)cr3cr2.pdb,
 		      cr3cr2.pcd ? "PCD" : "pcd",
